Move each state related logic to their own files
COAND-935
