+++ mbway/src/main/java/com/adyen/checkout/mbway/MBWayComponentState.kt
@@ -8,9 +8,15 @@
 
 package com.adyen.checkout.mbway
 
+import com.adyen.checkout.components.core.Amount
+import com.adyen.checkout.components.core.OrderRequest
 import com.adyen.checkout.components.core.PaymentComponentData
 import com.adyen.checkout.components.core.PaymentComponentState
+import com.adyen.checkout.components.core.internal.analytics.AnalyticsManager
+import com.adyen.checkout.components.core.internal.ui.model.transformer.FieldTransformerRegistry
 import com.adyen.checkout.components.core.paymentmethod.MBWayPaymentMethod
+import com.adyen.checkout.mbway.internal.ui.model.MBWayDelegateState
+import com.adyen.checkout.mbway.internal.ui.model.MBWayFieldId
 
 /**
  * Represents the state of [MBWayComponent].
@@ -20,3 +26,31 @@ data class MBWayComponentState(
     override val isInputValid: Boolean,
     override val isReady: Boolean
 ) : PaymentComponentState<MBWayPaymentMethod>
+
+internal fun MBWayDelegateState.toComponentState(
+    analyticsManager: AnalyticsManager,
+    fieldTransformerRegistry: FieldTransformerRegistry<MBWayFieldId>,
+    order: OrderRequest?,
+    amount: Amount?,
+): MBWayComponentState {
+    val paymentMethod = MBWayPaymentMethod(
+        type = MBWayPaymentMethod.PAYMENT_METHOD_TYPE,
+        checkoutAttemptId = analyticsManager.getCheckoutAttemptId(),
+        telephoneNumber = fieldTransformerRegistry.transform(
+            MBWayFieldId.LOCAL_PHONE_NUMBER,
+            localPhoneNumberFieldState.value,
+        ),
+    )
+
+    val paymentComponentData = PaymentComponentData(
+        paymentMethod = paymentMethod,
+        order = order,
+        amount = amount,
+    )
+
+    return MBWayComponentState(
+        data = paymentComponentData,
+        isInputValid = isValid,
+        isReady = true,
+    )
+}
